rename clone node file name
